requirements.txt
streamlit>=1.56.0
pandas>=3.0.0
scikit-learn>=1.8.0
scikit-learn-extra>=0.3.0
mysql-connector-python>=9.6.0
plotly>=6.6.0
matplotlib>=3.10.0
numpy>=2.4.0
